Bugfix for when initial scrollTop is > 0
When scroll position on page load is greater than zero (i.e. if you scroll to the middle of the page and reload it) the page would jump to the top on the first scroll, so you'd have to scroll all the way down again. It also happens if you click on anchors before the first scroll. Happens only once, but can be quite annoying.